requirements.txt
beautifulsoup4
requests
python-dotenv
pandas
streamlit
google-generativeai